+++ .env
@@ -1,4 +1,5 @@
 NODE_ENV='production',
+NODE_PATH=./src
 BASE_URL=null,
 LMS_BASE_URL=null,
 LOGIN_URL=null,
